move shared to shared/ui and create shared/designsystem gradle module
