+++ android/build.gradle.kts
@@ -53,7 +53,7 @@ kotlin {
 }
 
 dependencies {
-    implementation(projects.shared)
+    implementation(projects.shared.ui)
 
     implementation(libs.androidx.activity.compose)
     implementation(libs.androidx.splashscreen)
