+++ desktop/src/main/kotlin/app/chatty/desktop/Main.kt
@@ -1,8 +1,8 @@
 package app.chatty.desktop
 
-import app.chatty.shared.ContentView
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import app.chatty.shared.ui.ContentView
 
 fun main() {
     application {
