+++ shared/designsystem/src/androidMain/kotlin/app/chatty/shared/designsystem/theme/Theme.android.kt
@@ -1,0 +1,23 @@
+package app.chatty.shared.designsystem.theme
+
+import android.os.Build
+import androidx.annotation.ChecksSdkIntAtLeast
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+
+@ChecksSdkIntAtLeast(Build.VERSION_CODES.S)
+actual fun supportDynamicColor(): Boolean = Build.VERSION.SDK_INT >= Build.VERSION_CODES.S
+
+@Composable
+internal actual fun colorScheme(
+    darkTheme: Boolean,
+    dynamicColor: Boolean,
+): ColorScheme = if (dynamicColor && supportDynamicColor()) {
+    val context = LocalContext.current
+    if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
+} else {
+    if (darkTheme) DarkColorScheme else LightColorScheme
+}
