+++ shared/designsystem/src/commonMain/kotlin/app/chatty/shared/designsystem/theme/Type.kt
@@ -1,0 +1,5 @@
+package app.chatty.shared.designsystem.theme
+
+import androidx.compose.material3.Typography
+
+internal val Typography = Typography()
