--- shared/src/commonMain/kotlin/app/chatty/shared/theme/Type.kt
@@ -1,5 +1,0 @@
-package app.chatty.shared.theme
-
-import androidx.compose.material3.Typography
-
-internal val Typography = Typography()
