--- shared/src/jvmMain/kotlin/app/chatty/shared/theme/Theme.jvm.kt
@@ -1,20 +1,0 @@
-package app.chatty.shared.theme
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-
-@Composable
-actual fun ChattyTheme(
-    darkTheme: Boolean,
-    dynamicColor: Boolean,
-    content: @Composable () -> Unit,
-) {
-    MaterialTheme(
-        colorScheme = if (darkTheme) DarkColorScheme else LightColorScheme,
-        shapes = Shapes,
-        typography = Typography,
-        content = content
-    )
-}
-
-actual fun supportDynamicColor(): Boolean = false
