+++ shared/ui/build.gradle.kts
@@ -1,0 +1,32 @@
+plugins {
+    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.jetbrains.compose)
+    alias(libs.plugins.kotlin.compose.compiler)
+}
+
+kotlin {
+    androidTarget()
+    jvm()
+
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.shared.designsystem)
+
+            implementation(compose.runtime)
+            implementation(compose.foundation)
+            implementation(compose.material3)
+            implementation(compose.ui)
+            implementation(compose.components.resources)
+            implementation(compose.components.uiToolingPreview)
+        }
+    }
+
+    jvmToolchain(11)
+}
+
+android {
+    namespace = "app.chatty.shared.ui"
+    compileSdk = 34
+    defaultConfig.minSdk = 21
+}
