+++ shared/ui/src/commonMain/kotlin/app/chatty/shared/ui/ContentView.kt
@@ -1,0 +1,29 @@
+package app.chatty.shared.ui
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import app.chatty.shared.designsystem.theme.ChattyTheme
+import org.jetbrains.compose.ui.tooling.preview.Preview
+
+@Composable
+@Preview
+fun ContentView() {
+    ChattyTheme {
+        Scaffold {
+            Box(
+                contentAlignment = Alignment.Center,
+                modifier = Modifier
+                    .padding(it)
+                    .fillMaxSize()
+            ) {
+                Text(text = "Hi, Mom!")
+            }
+        }
+    }
+}
